+++ src/main/java/ir/mapsa/deathcorridorgame/rifle/SniperRifle.java
@@ -9,7 +9,7 @@ public class SniperRifle extends Rifle{
     private boolean isZoom;
 
     public SniperRifle(Bullet bullet, boolean isZoom) {
-        super(0.60f, 20, Accuracy.HIGH, bullet);
+        super(0.60, 20, Accuracy.HIGH, bullet);
         this.isZoom = isZoom;
         checkZoom();
     }
@@ -20,9 +20,9 @@ public class SniperRifle extends Rifle{
 
     private void checkZoom() {
         if(isZoom){
-            float random = 1;
+            double random = 1;
             try {
-                random = Helper.showRandomPercent(0.05f, 0.15f, new Random());
+                random = Helper.showRandomPercent(0.05, 0.15, new Random());
             }catch (IllegalArgumentException e){
                 System.err.println("Can't calculate random about zoom of rifle because: " + e.getMessage());
             }
